533 Move pas-form optionsets into their own file
