+++ client/app/models/pas-form.js
@@ -1,62 +1,5 @@
 import Model, { belongsTo, hasMany, attr } from '@ember-data/model';
 
-export const YES_NO_UNSURE_OPTIONSET = {
-  YES: {
-    code: 717170000,
-    label: 'Yes',
-  },
-  NO: {
-    code: 717170001,
-    label: 'No',
-  },
-  UNSURE: {
-    code: 717170002,
-    label: 'Unsure at this time',
-  },
-};
-
-export const DCPLEGALSTREETFRONTAGE_OPTIONSET = {
-  MAPPED_AND_BUILT: {
-    code: 717170000,
-    label: 'Mapped and Built',
-  },
-  PRIVATE_ROAD: {
-    code: 717170001,
-    label: 'Private Road',
-  },
-  RECORD_STREET: {
-    code: 717170002,
-    label: 'Record Street',
-  },
-  CORPORATION_COUNSEL_OPINION: {
-    code: 717170003,
-    label: 'Corporation Counsel Opinion',
-  },
-  MAPPED_BUT_NOT_BUILT: {
-    code: 717170004,
-    label: 'Mapped but not Built',
-  },
-};
-
-export const DCPHOUSINGUNITTYPE_OPTIONSET = {
-  CITY: {
-    code: 717170000,
-    label: 'City',
-  },
-  STATE: {
-    code: 717170000,
-    label: 'State',
-  },
-  FEDERAL: {
-    code: 717170000,
-    label: 'Federal',
-  },
-  OTHER: {
-    code: 717170000,
-    label: 'Other',
-  },
-};
-
 export default class PasFormModel extends Model {
   @belongsTo('package', { async: false })
   package;
